requirements_streamlit.txt
streamlit==1.28.0
tensorflow==2.15.0
numpy==1.24.3
pillow==10.0.0
opencv-python==4.8.1.78
matplotlib==3.7.2
streamlit-drawable-canvas==0.9.3
